Fixing save data and save plot

+++ R/AlradSpectra.R
@@ -114,17 +114,21 @@ AlradSpectra <- function() {
                                                                 ylab = ylab)
                                                         }
   # Export plot as png graphics file
-  fsaveplot    <- function(w, h,...)  {fname <- paste0(gfile(text="Save File", type="save", initialfilename="Plot", container=window,
-                                                             filter=c("Portable Network Graphics (.png)"="png")),".png")
-                                                dev.copy(png, fname, width=w, height=h, res=100, antialias = "cleartype")
-                                                dev.off() #Close graphics device
+  fsaveplot    <- function(w, h,...)  {fdialog <- gfile("Save File", type="save", initialfilename="Plot", container=window,
+                                                        filter=c("Portable Network Graphics (.png)"="png"))
+                                       if(!(is.na(fdialog))) #If fdialog is not equal to NA, keep running
+                                         fname <- paste0(fdialog,".png")
+                                         dev.copy(png, fname, width=w, height=h, res=100, antialias = "cleartype")
+                                         dev.off() #Close graphics device
                                        }
   # Export preprocessed spectra as csv file
-  fsavedata    <- function(h, ...)    {fname      <- paste0(gfile(text="Save File", type="save", initialfilename="Data", container=window,
-                                                                  filter=c("Comma Separated Values (.csv)"="csv")),".csv")
-                                       spectrum   <- seq(spectra.start.column, spectra.end.column)
-                                       exportdata <- cbind(alldata[,-spectrum], h)
-                                       write.csv(exportdata, row.names = FALSE, file = fname)
+  fsavedata    <- function(h, ...)    {fdialog <- gfile("Save File", type="save", initialfilename="Output", container=window,
+                                                        filter=c("Portable Network Graphics (.png)"="png"))
+                                       if(!(is.na(fdialog))) #If fdialog is not equal to NA, keep running
+                                         fname      <- paste0(fdialog,".csv")
+                                         spectrum   <- seq(spectra.start.column, spectra.end.column)
+                                         exportdata <- cbind(alldata[,-spectrum], h)
+                                         write.csv(exportdata, row.names = FALSE, file = fname)
                                        }
   # Adds preprocessing to combobox in Model tab only if it is not already there
   faddtolist   <- function(h, ...)    {present <- is.element(h, dataset)
